Remove dup `declare_tests!` macros (#1063)
* Remove dup declare test macros

* Introduce review suggestions

* Change mod names

* Add review suggestion

* Update integration test name

* Fix naming

* Fix features

* Fix: use mock prover in mock functions
